backend/requirements.txt
django
djangorestframework
Pillow
djangorestframework-simplejwt
django-cors-headers
python-dotenv
gunicorn
pdfkit